wip: move words hooks into game state provider
