+++ app/src/Components/Layout/GameStateProvider/useCurrentGameState.ts
@@ -1,7 +1,8 @@
 import { useCallback, useMemo } from 'react'
 
 import { useLocalStorage } from '../../../hooks/useLocalStorage'
-import { type WordListLength } from '../../../hooks/words/useWordList'
+
+import { type WordListLength } from './words/useWordList'
 
 type GameState = {
     guesses: Array<string>
